client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "langgraph>=0.2.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
